+++ src/Account/Dto/AccountCreateRequest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Shellrent\Arubapec\Account\Dto;
 
 use Shellrent\Arubapec\Shared\Dto\ContractData;
+use Shellrent\Arubapec\Shared\Dto\OwnerId;
 use Shellrent\Arubapec\Shared\Dto\RenewalData;
 
 final class AccountCreateRequest